add full type definitions to Response
