--- src/index.d.ts
@@ -1,5 +1,0 @@
-declare namespace Express {
-    export interface Response {
-        sse(data: string): void;
-    }
-}
