+++ src/response.ts
@@ -1,0 +1,5 @@
+import { Response as ExpressResponse } from "express";
+
+export interface Response extends ExpressResponse {
+    sse(data: string): void;
+}
